requirements.txt
asgiref==3.7.2
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
distlib==0.3.6
Django==4.2.3
django-allauth==0.33.0
django-cors-headers==4.2.0
django-rest-auth==0.9.2
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.12.0
idna==2.6
oauthlib==2.0.6
pipenv==2023.5.19
pipreqs==0.4.13
platformdirs==3.5.1
psycopg==3.1.9
psycopg-binary==3.1.9
python3-openid==3.1.0
pytz==2023.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.22
virtualenv==20.23.0
virtualenv-clone==0.5.7
yarg==0.1.9
